requirements.txt
cssmin==0.2.0
Django==1.9.5
django-grappelli==2.8.1
django-appconf==1.0.1
django-compressor==2.0
django-constance==1.1
django-extensions==1.6.7
django-ipware==1.1.0
django-parsley==0.5
django-picklefield==0.3.2
django-post-office==2.0.7
django-reversion==2.0.5
easy-thumbnails==2.2
guppy==0.1.10
jsmin==2.1.2
jsonfield==1.0.3
Pillow==3.3.0
psycopg2==2.6.1
pytils==0.3
pytz==2015.4
rcssmin==1.0.6
requests==2.7.0
rjsmin==1.0.12
six==1.9.0
sqlparse==0.1.15
tablib==0.10.0
Werkzeug==0.11.10
django-cities==0.4.2
django-parler==1.6.5
openexchangerates==0.1.1
django-currencies==0.4.0
pinax-stripe==3.3.0
django-cities-light==3.3.0
